+++ src/main/java/com/telepathicgrunt/the_bumblezone/Bumblezone.java
@@ -23,6 +23,7 @@ import com.telepathicgrunt.the_bumblezone.modCompat.ModdedBeesBeesSpawning;
 import com.telepathicgrunt.the_bumblezone.modCompat.PotionOfBeesBeeSplashPotionProjectile;
 import com.telepathicgrunt.the_bumblezone.surfacebuilders.BzSurfaceBuilders;
 import com.telepathicgrunt.the_bumblezone.tags.BZBlockTags;
+import com.telepathicgrunt.the_bumblezone.tags.BZItemTags;
 import com.telepathicgrunt.the_bumblezone.utils.ConfigHelper;
 import net.minecraft.block.Block;
 import net.minecraft.entity.Entity;
@@ -58,6 +59,7 @@ public class Bumblezone{
         IEventBus forgeBus = MinecraftForge.EVENT_BUS;
         IEventBus modEventBus = FMLJavaModLoadingContext.get().getModEventBus();
         BZBlockTags.tagInit(); // Done extra early as some features needs the tag wrapper.
+        BZItemTags.tagInit();
 
         //Events
         //forgeBus.addListener(BzDimension::biomeModification);