+++ src/main/java/com/telepathicgrunt/the_bumblezone/entities/mobs/HoneySlimeEntity.java
@@ -84,7 +84,7 @@ public class HoneySlimeEntity extends AnimalEntity implements IAngerable, IMob {
 
    @Override
    public ILivingEntityData onInitialSpawn(IServerWorld worldIn, DifficultyInstance difficultyIn, SpawnReason reason, ILivingEntityData spawnDataIn, CompoundNBT dataTag) {
-      this.setSlimeSize(2, true);
+      this.setSlimeSize(this.isChild() ? 1 : 2, true);
       return super.onInitialSpawn(worldIn, difficultyIn, reason, spawnDataIn, dataTag);
    }