+++ src/main/java/com/telepathicgrunt/the_bumblezone/tags/BZItemTags.java
@@ -1,0 +1,15 @@
+package com.telepathicgrunt.the_bumblezone.tags;
+
+import com.telepathicgrunt.the_bumblezone.Bumblezone;
+import net.minecraft.block.Block;
+import net.minecraft.item.Item;
+import net.minecraft.tags.BlockTags;
+import net.minecraft.tags.ITag;
+import net.minecraft.tags.ItemTags;
+
+public class BZItemTags {
+    // All tag wrappers need to be made at mod init.
+    public static void tagInit(){}
+
+    public static final ITag.INamedTag<Item> TURN_SLIME_TO_HONEY_SLIME = ItemTags.makeWrapperTag(Bumblezone.MODID+":turn_slime_to_honey_slime");
+}